Organize Try and Buy by country and update UI
