+++ frontend/pages/dashboard.jsx
@@ -26,36 +26,11 @@ function DashboardPage() {
     return () => { cancelled = true; };
   }, [router.query.code]);
 
-  const gtLink = code ? `/galaxy-try/${code}` : "/galaxy-try";
   const devLink = code ? `/c/${code}/devices` : "/devices";
-  const tryBuyLink = "/try-and-buy";
+  const tryBuyLink = code ? `/try-and-buy/${code}` : "/try-and-buy";
 
   return (
-    <div className="grid grid-cols-3 h-screen bg-gradient-to-br from-samsung-blue to-black">
-      <Link href={gtLink} className="relative block">
-        <div
-          className="absolute inset-0 bg-center bg-cover"
-          style={{ backgroundImage: "url('/Background%20galaxytry.jpg')" }}
-        />
-        <div className="absolute inset-0 bg-black/40" />
-        <div className="relative flex h-full w-full items-center justify-center">
-          <span className="text-white text-4xl md:text-5xl font-bold">
-            GALAXY TRY
-          </span>
-        </div>
-      </Link>
-      <Link href={devLink} className="relative block">
-        <div
-          className="absolute inset-0 bg-center bg-cover"
-          style={{ backgroundImage: "url('/Background%20foldables.jpg')" }}
-        />
-        <div className="absolute inset-0 bg-black/40" />
-        <div className="relative flex h-full w-full items-center justify-center">
-          <span className="text-white text-4xl md:text-5xl font-bold">
-            DEVICES
-          </span>
-        </div>
-      </Link>
+    <div className="grid grid-cols-2 h-screen bg-gradient-to-br from-samsung-blue to-black">
       <Link href={tryBuyLink} className="relative block">
         <div
           className="absolute inset-0 bg-center bg-cover"
@@ -68,6 +43,18 @@ function DashboardPage() {
         <div className="relative flex h-full w-full items-center justify-center">
           <span className="text-white text-4xl md:text-5xl font-bold">
             TRY & BUY
+          </span>
+        </div>
+      </Link>
+      <Link href={devLink} className="relative block">
+        <div
+          className="absolute inset-0 bg-center bg-cover"
+          style={{ backgroundImage: "url('/Background%20foldables.jpg')" }}
+        />
+        <div className="absolute inset-0 bg-black/40" />
+        <div className="relative flex h-full w-full items-center justify-center">
+          <span className="text-white text-4xl md:text-5xl font-bold">
+            DEVICES
           </span>
         </div>
       </Link>
