+++ frontend/pages/try-and-buy/hr.tsx
@@ -1,0 +1,2 @@
+import TryAndBuyPage from "../../components/TryAndBuyPage";
+export default TryAndBuyPage;
