+++ frontend/pages/try-and-buy/rs.tsx
@@ -1,0 +1,2 @@
+import TryAndBuyPage from "../../components/TryAndBuyPage";
+export default TryAndBuyPage;
