+++ frontend/pages/try-and-buy/si.tsx
@@ -1,0 +1,2 @@
+import TryAndBuyPage from "../../components/TryAndBuyPage";
+export default TryAndBuyPage;
